refactor(commands): extract plugin commands into plugin.rs
